# Optimize insertions with `multiEntry`