This is a slight optimization on top of #52, to help with #44. The improvement to the original benchmark is (measuring the mean with `hyperfine`):

| Before | After | Delta |
| --- | --- | --- |
| 4.444 s ±  0.117 s | 3.985 s ±  0.098 s | **459 ms (-10%)** |

We can show the benefits more clearly by slightly tweaking ([dae0981](https://github.com/nolanlawson/fakeIndexedDB/commit/dae098173556e2a49cb02d6f5e804caaeaef8181)) the benchmark to have more duplicate `multiEntry` keys. Then the result is (`hyperfine` with 25 iterations):

| Before | After | Delta |
| --- | --- | --- |
| 419.6 ms ±  18.0 ms | 110.2 ms ±   5.5 ms | **309 ms (-74%)** |

The reason this is faster is because, instead of doing a binary search on the keys and then a linear search on the values, we are doing a binary search on both. If you have a lot of duplicate `multiEntry` keys, then you end up with a high ratio of values (i.e. record primary keys) to keys, so a linear search would be slow.

I think that if we want to get much faster than this, then we'd need to implement a BTree or similar. A lot of the remaining time is eaten up by `Array.prototype.splice()`.